+++ CHANGELOG.md
@@ -58,6 +58,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Card_page : Big loyalty card render problem
 - Loyalty Cards Page : User can now scan different barcode types
 - Loyalty Cards Page : Cancel a card creation don't create cards anymore
+- Widget loyaltycards : Loyalty card name too short
 
 ### Security